CMake: generate linker script fragments for segments
* Animations
* Levels
* Dynamic models
* Strings

Started to hook this up to the linker. There are still some errors to fix.

+++ src/util/string.c
@@ -1,7 +1,7 @@
 #include "string.h"
 
 size_t strLength(char* s) {
-    int size = 0;
+    size_t size = 0;
 
     while (*s++ != '\0') {
         ++size;
